- Send RouteRewriteMessage

+++ src/mesh/node_test.go
@@ -3,7 +3,9 @@ package mesh
 import(
 	"time"
 	"reflect"
-	"testing")
+	"testing"
+	//"fmt"
+	)
 
 import (
 	"github.com/stretchr/testify/assert"
@@ -52,7 +54,6 @@ func TestEstablishRoutes(t *testing.T) {
 			OperationReply{OperationMessage{Message{0}, establish_msg.MsgId}, true, ""}, 
 			55, "secret_abc"}
 	}
-
 	// Establish 3 -> 4
 	{
 		outgoing := <- node.MessagesOut
@@ -67,7 +68,6 @@ func TestEstablishRoutes(t *testing.T) {
 			OperationReply{OperationMessage{Message{55}, establish_msg.MsgId}, true, ""}, 
 			120, "secret_xyz"}
 	}
-/*
 	// Set rewrite 2 -> 3
 	{
 		outgoing := <- node.MessagesOut
@@ -76,18 +76,18 @@ func TestEstablishRoutes(t *testing.T) {
 		assert.Equal(t, reflect.TypeOf(msg), reflect.TypeOf(RouteRewriteMessage{}))
 		rewrite_msg := msg.(RouteRewriteMessage)
 		assert.Equal(t, rewrite_msg.Secret, "secret_abc")
-		assert.Equal(t, rewrite_msg.RewriteSendId, 120)
+		assert.Equal(t, rewrite_msg.RewriteSendId, uint32(120))
 
 		// Reply
 		node.MessagesIn <- OperationReply{OperationMessage{Message{0}, rewrite_msg.MsgId}, true, ""}
 	}
-	*/
-
 	route_established := <- established_ch
 
 	assert.Equal(t, route_established.SendId, uint32(55))
 	assert.Equal(t, route_established.ConnectedPeer, test_key2)
 }
+
+// Establish route test, rewrite test
 
 // Retransmit, stop retransmitting
 // Timeout
